refactor: car names 유효성 검사 함수 리팩토링
