+++ src/game/utils/isValidCarNamesInput.js
@@ -1,8 +1,14 @@
-import isValidCarNames from './isValidCarNames.js';
+import hasDuplicate from '../../utils/hasDuplicate.js';
 import toCarNameList from './toCarNameList.js';
+import isValidCarName from './isValidCarName.js';
 
 function isValidCarNamesInput(input) {
-  return isValidCarNames(toCarNameList(input));
+  const carNames = toCarNameList(input);
+
+  return (
+    !hasDuplicate(carNames) &&
+    carNames.every((carName) => isValidCarName(carName))
+  );
 }
 
 export default isValidCarNamesInput;
