+++ src/game/utils/isValidRacingCountInput.js
@@ -1,4 +1,4 @@
-import { VALIDE_CONDITION_RACING_COUNT } from './const.js';
+import { VALIDE_CONDITION_RACING_COUNT } from '../const.js';
 
 function isValidRacingCountInput(input) {
   const count = Number(input);
